feat: Can now create RaytracingPSO. An abstraction for RT PSO creation.

+++ DX12RadianceCascades/src/RadianceCascadesManager.cpp
@@ -23,7 +23,7 @@ void RadianceCascadesManager::Init(float _rayLength0, float _raysPerProbe0, floa
 
 	uint32_t cascadeCount = 6;
 	float probeSpacing = 2.0f;
-	const uint32_t probeCountPerDim0 = (uint32_t)Math::AlignPowerOfTwo<float>(1024.0f / probeSpacing);
+	const uint32_t probeCountPerDim0 = (uint32_t)Math::AlignPowerOfTwo<uint32_t>((uint32_t)(1024.0f / probeSpacing));
 
 	if (m_cascadeIntervals.size() < cascadeCount)
 	{
